+++ ydb/public/sdk/python/ydb/connection.py
@@ -146,6 +146,7 @@ def _construct_metadata(driver_config, settings):
             metadata.append((YDB_TRACE_ID_HEADER, settings.trace_id))
         if settings.request_type is not None:
             metadata.append((YDB_REQUEST_TYPE_HEADER, settings.request_type))
+        metadata.extend(getattr(settings, "headers", []))
 
     metadata.append(_utilities.x_ydb_sdk_build_info_header())
     return metadata